renamed models directory containing redux types
