+++ src/redux/reducers/coffeesReducer.ts
@@ -2,7 +2,7 @@ import { CoffeeItem } from '../../models/modelResource';
 import {
   ReduxDispatch,
   ReduxActions,
-} from '../../models/reduxSaga/reduxResource';
+} from '../../models/redux/reduxResource';
 
 // coffeesReducer contains all the coffees displayed on a user's dashboard
 // also contains search results
