+++ src/redux/reducers/flavorsReducer.ts
@@ -2,7 +2,7 @@ import { Flavors } from '../../models/modelResource';
 import {
   ReduxActions,
   ReduxDispatch,
-} from '../../models/reduxSaga/reduxResource';
+} from '../../models/redux/reduxResource';
 
 // flavorsReducer contains list of flavors, which are displayed as Chips
 // in various places around the app
