+++ src/redux/reducers/sharedCoffeesReducer.ts
@@ -2,7 +2,7 @@ import { SharedCoffees } from '../../models/modelResource';
 import {
   ReduxActions,
   ReduxDispatch,
-} from '../../models/reduxSaga/reduxResource';
+} from '../../models/redux/reduxResource';
 
 // sharedCoffeesReducer contains any entries of shared coffees
 // sent by other users, this is checked in UseEffect() on Dashboard
