+++ src/redux/sagas/flavorsSaga.ts
@@ -3,8 +3,8 @@ import { put, takeEvery, call } from 'redux-saga/effects';
 import {
   SagaActions,
   SagaGeneratorReturn,
-} from '../../models/reduxSaga/sagaResource';
-import { ReduxActions } from '../../models/reduxSaga/reduxResource';
+} from '../../models/redux/sagaResource';
+import { ReduxActions } from '../../models/redux/reduxResource';
 import { Flavors } from '../../models/modelResource';
 
 // Fetches list of broad palette flavors used throughout the app
